+++ src/test/java/com/synopsys/defensics/api/ApiServiceTest.java
@@ -17,17 +17,22 @@
 package com.synopsys.defensics.api;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.fail;
 
+import com.synopsys.defensics.apiserver.client.DefensicsApiV2Client;
 import com.synopsys.defensics.apiserver.model.Run;
 import com.synopsys.defensics.apiserver.model.RunState;
 import com.synopsys.defensics.client.DefensicsRequestException;
 import com.synopsys.defensics.jenkins.test.utils.DefensicsMockServer;
 import hudson.FilePath;
 import java.io.File;
+import java.net.URI;
+import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -127,5 +132,42 @@ public class ApiServiceTest {
     api.saveResults(run, resultFolder);
     assertThat(resultFolder.exists(), is(equalTo(true)));
     assertThat(resultFolder.child("report.html").exists(), is(equalTo(true)));
+  }
+
+  /**
+   * Test following things:
+   *
+   * 1) Client timeout setting works
+   * 2) Exception thrown from exceeding timeouts is DefensicsRequestException instead of
+   *   InterruptedException so that'd make build to FAIL instead of ABORTED.
+   */
+  @Test
+  public void testClientTimeoutHandling() {
+    final URI apiBaseUri = URI.create(DEFENSICS_URL);
+    final DefensicsApiV2Client clientWithShortTimeout = new DefensicsApiV2Client(
+        apiBaseUri.resolve("/api/v2"),
+        TOKEN,
+        builder -> builder.readTimeout(1, TimeUnit.MILLISECONDS)
+    );
+    final ApiService apiService = new ApiService(clientWithShortTimeout, apiBaseUri);
+    final DefensicsRequestException defensicsRequestException = assertThrows(
+        DefensicsRequestException.class,
+        // Mock server is configured to have 5 ms delay for create-run request
+        () -> apiService.createNewRun()
+    );
+
+    System.out.println(defensicsRequestException.getMessage());
+    assertThat(
+        defensicsRequestException.getMessage(),
+        containsString("Could not create test run")
+    );
+    // timeout wording varies a bit between test environments
+    assertThat(
+        defensicsRequestException.getMessage(),
+        anyOf(
+            containsString("timeout"),
+            containsString("Read timed out")
+        )
+    );
   }
 }